requirements.txt
streamlit
google-genai
streamlit-image-select